--- begin-react/src/App.js
@@ -1,13 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import Detail from './Detail.js';
-
-function App() {
-  return (
-    <div className="App">
-      <Detail></Detail>
-    </div>
-  );
-}
-
-export default App;
