+++ src/Detail.css
@@ -115,7 +115,7 @@
     display: flex;
     flex-wrap: wrap;
     justify-content: center;
-   
+    margin-bottom: 100px;
     gap: 30px;
   }
   
@@ -144,4 +144,61 @@
     margin-right:3px;
   }
   
-
+  #comment-box {
+    background-color: #f2f2f2;
+    padding: 20px;
+    border-radius: 5px;
+    margin-bottom: 20px;
+  }
+  
+  #comment-input {
+    width: 100%;
+    height: 80px;
+    padding: 120px;
+    font-size: 16px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    resize: none;
+    margin-bottom: 10px;
+  }
+  
+  #comment-submit {
+    background-color: #4CAF50;
+    color: white;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+  }
+  
+  #comment-submit:hover {
+    background-color: #45a049;
+  }
+  
+  #comment-list {
+    margin-top: 20px;
+  }
+  
+  #comment-list li {
+    background-color: white;
+    padding: 10px;
+    border-radius: 5px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    margin-bottom: 10px;
+  }
+  
+  #comment-list li p {
+    margin: 0;
+    font-size: 14px;
+  }
+  
+  #comment-list li p:first-child {
+    font-weight: bold;
+  }
+  
+  #comment-list li p:last-child {
+    color: #888;
+    font-size: 12px;
+  }
+
